requirements.txt
pandas~=1.5.3
scikit-learn~=1.2.2
autogluon~=0.8.2
matplotlib==3.4.3
joblib~=1.4.2
python-dotenv
streamlit
langchain-openai